+++ src/main/java/com/atlauncher/mclauncher/MCLauncher.java
@@ -164,6 +164,8 @@ public class MCLauncher {
         int initialMemory = Optional.ofNullable(instance.launcher.initialMemory).orElse(App.settings.initialMemory);
         int maximumMemory = Optional.ofNullable(instance.launcher.maximumMemory).orElse(App.settings.maximumMemory);
         int permGen = Optional.ofNullable(instance.launcher.permGen).orElse(App.settings.metaspace);
+        boolean enableLog4jExploitFix = Optional.ofNullable(instance.launcher.enableLog4jExploitFix)
+                .orElse(App.settings.enableLog4jExploitFix);
         String javaArguments = Optional.ofNullable(instance.launcher.javaArguments).orElse(App.settings.javaParameters);
         String javaPath = instance.getJavaPath();
 
@@ -326,6 +328,11 @@ public class MCLauncher {
         if (OS.isWindows() && !arguments
                 .contains("-XX:HeapDumpPath=MojangTricksIntelDriversForPerformance_javaw.exe_minecraft.exe.heapdump")) {
             arguments.add("-XX:HeapDumpPath=MojangTricksIntelDriversForPerformance_javaw.exe_minecraft.exe.heapdump");
+        }
+
+        // add -Dlog4j2.formatMsgNoLookups=true if not there already (Log4J exploit fix)
+        if (enableLog4jExploitFix && !arguments.contains("-Dlog4j2.formatMsgNoLookups=true")) {
+            arguments.add("-Dlog4j2.formatMsgNoLookups=true");
         }
 
         // if there's no -Djava.library.path already, then add it (for older versions)