+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### New Features
 - Show dialog to confirm users action when creating a server without any instances
+- Add JVM argument to prevent Log4J zero day exploit
 
 ### Fixes
 - Issue with CurseForge packs using forge-recommended or forge-latest failing to install