+++ CHANGELOG.md
@@ -3,10 +3,9 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.10.5
+## 3.4.10.6
 
 ### New Features
-- Download latest Minecraft manifest on launch
 
 ### Fixes